Module poc_offtake_parser::electricity_offtake_parser

Elexon Files

The file is a pipe seperated list of offtake values for a given day in a specfic location. A single line looks like so:

DPP|1|20230401|.0000491

Which is identifer|location|date|value

These are grouped into Profile Class Groups with lines:

PFC|1

Where the number is the profile class identifer

Re-exports

pub use parser::parse_flow_file;

Modules

parser 🔒

Structs

HeaderLine

OfftakeFile

OfftakeRow

ProfileGroups

TrailerLine

WrappedFile

Enums

Location

ProfileClass