Module poc_offtake_parser::electricity_offtake_parser::parser

Functions

date 🔒

Parse the date in the form YearMonthDay

date_time 🔒

Parse the date and time in the form YearMonthDayHourMinuteSecond

header 🔒

Gets the header from an electricity flow file. Since we don’t care about most of the details we only consume the date from the file.

offtake_file 🔒

A looping parser over our offtake lines with the PFC grouping. These groups are seperated by PFC and end when we reach the trailer with ZPT

offtake_group_parser 🔒

Looping parser for offtake lines starting with DPP, we use the start of each line as the seperator. We know we have finished when we reach a new profile group starting with PFC or the end of the file with ZPT

offtake_group_terminator 🔒

The end of set of offtake lines could either be a new profile group or the trailer

offtake_line 🔒

Parse a single offtake line

parse_flow_file

The main parsing function that collects our other parsers. First get the header line. Then get all of the offtake lines and finally grab the trailer.

profile_class 🔒

From the profile class line grab the profile class number ie PFC|1 => ProfileClass:Class1 This can fail if the line is malformed or if the number isnt a know profile class.

profile_class_group 🔒

Parser the profile class with its offtake lines. It expects a group of the form:

trailer 🔒

parse the trailer line.