Module poc_offtake_parser::gas_offtake_parser::parser

Xoserve Files

Xoserve provide two files that must be read to build the seasonality data.

SNCWV

The Seasonal Normal Composite Weather Variables file is a comma separated text file like so

"01/10/2022", 11.08, 10.65, 11.82, 11.90, 12.14, 12.09, 12.43, 12.82, 14.37, 14.32, 13.78, 14.21, 12.96
"02/10/2022", 10.95, 10.52, 11.72, 11.78, 12.03, 11.97, 12.33, 12.74, 14.26, 14.20, 13.68, 14.13, 12.89
"03/10/2022", 10.87, 10.41, 11.62, 11.68, 11.93, 11.88, 12.24, 12.66, 14.15, 14.09, 13.60, 14.06, 12.81
"04/10/2022", 10.79, 10.31, 11.52, 11.59, 11.85, 11.80, 12.15, 12.58, 14.06, 13.99, 13.51, 13.98, 12.74

Each Local Distribution Zone (LDZ) has is own value in a column.

ALDAF

The Annual Load Profile and Daily Adjustment Factor file is also comma separated

"EA:E2201BND",01/10/2022,0.589078,-0.238091
"EA:E2201BNI",01/10/2022,0.543703,-0.191565
"EA:E2201BPD",01/10/2022,0.595346,-0.222440
"EA:E2201BPI",01/10/2022,0.543703,-0.191565
"EA:E2202BND",01/10/2022,0.712085,-0.141062

Here the values are grouped by LDZ and also the End User Catergory (EUC). There are multiple EUCs but we only provide energy currently for domestic and small businesses so only use the rows E2201BND and E2202BND

Functions

alpdaf_file 🔒

alpdaf_row 🔒

date 🔒

parse a date surrounded by quotation marks

end_user_category 🔒

raw_date 🔒

Parse a date in the form day/month/year

scwm_file 🔒

scwv_row 🔒